+++ src/app/upload/components/parseRecipe.ts
